Move EnsureAuthorized functions from user client to auth client
