+++ auth/auth.go
@@ -22,6 +22,9 @@ type Client interface {
 type ExternalAccessor interface {
 	ServerSessionToken() (string, error)
 	ValidateSessionToken(ctx context.Context, token string) (request.Details, error)
+	EnsureAuthorized(ctx context.Context) error
+	EnsureAuthorizedService(ctx context.Context) error
+	EnsureAuthorizedUser(ctx context.Context, targetUserID string, permission string) (string, error)
 }
 
 type contextKey string
