+++ auth/client/external.go
@@ -11,6 +11,7 @@ import (
 	"github.com/tidepool-org/platform/config"
 	"github.com/tidepool-org/platform/errors"
 	"github.com/tidepool-org/platform/log"
+	"github.com/tidepool-org/platform/permission"
 	"github.com/tidepool-org/platform/platform"
 	"github.com/tidepool-org/platform/request"
 )
@@ -182,6 +183,72 @@ func (e *External) ValidateSessionToken(ctx context.Context, token string) (requ
 	}
 
 	return request.NewDetails(request.MethodSessionToken, result.UserID, token), nil
+}
+
+func (e *External) EnsureAuthorized(ctx context.Context) error {
+	if ctx == nil {
+		return errors.New("context is missing")
+	}
+
+	if details := request.DetailsFromContext(ctx); details != nil {
+		return nil
+	}
+
+	return request.ErrorUnauthorized()
+}
+
+func (e *External) EnsureAuthorizedService(ctx context.Context) error {
+	if ctx == nil {
+		return errors.New("context is missing")
+	}
+
+	if details := request.DetailsFromContext(ctx); details != nil {
+		if details.IsService() {
+			return nil
+		}
+	}
+
+	return request.ErrorUnauthorized()
+}
+
+func (e *External) EnsureAuthorizedUser(ctx context.Context, targetUserID string, authorizedPermission string) (string, error) {
+	if ctx == nil {
+		return "", errors.New("context is missing")
+	}
+	if targetUserID == "" {
+		return "", errors.New("target user id is missing")
+	}
+	if authorizedPermission == "" {
+		return "", errors.New("authorized permission is missing")
+	}
+
+	if details := request.DetailsFromContext(ctx); details != nil {
+		if details.IsService() {
+			return "", nil
+		}
+
+		authenticatedUserID := details.UserID()
+		if authenticatedUserID == targetUserID {
+			if authorizedPermission != permission.Custodian {
+				return authenticatedUserID, nil
+			}
+		} else {
+			url := e.client.ConstructURL("access", targetUserID, authenticatedUserID)
+			permissions := permission.Permissions{}
+			if err := e.client.RequestData(ctx, "GET", url, nil, nil, &permissions); err != nil {
+				if !request.IsErrorResourceNotFound(err) {
+					return "", errors.Wrap(err, "unable to get user permissions")
+				}
+			} else {
+				permissions = permission.FixOwnerPermissions(permissions)
+				if _, ok := permissions[authorizedPermission]; ok {
+					return authenticatedUserID, nil
+				}
+			}
+		}
+	}
+
+	return "", request.ErrorUnauthorized()
 }
 
 func (e *External) timeoutServerSessionToken(serverSessionTokenTimeout time.Duration) time.Duration {
