+++ data/source/service/test/client_provider.go
@@ -1,23 +1,39 @@
 package test
 
 import (
+	"github.com/tidepool-org/platform/auth"
 	dataSourceStoreStructured "github.com/tidepool-org/platform/data/source/store/structured"
-	"github.com/tidepool-org/platform/user"
 )
 
 type ClientProvider struct {
+	AuthClientInvocations                int
+	AuthClientStub                       func() auth.Client
+	AuthClientOutputs                    []auth.Client
+	AuthClientOutput                     *auth.Client
 	DataSourceStructuredStoreInvocations int
 	DataSourceStructuredStoreStub        func() dataSourceStoreStructured.Store
 	DataSourceStructuredStoreOutputs     []dataSourceStoreStructured.Store
 	DataSourceStructuredStoreOutput      *dataSourceStoreStructured.Store
-	UserClientInvocations                int
-	UserClientStub                       func() user.Client
-	UserClientOutputs                    []user.Client
-	UserClientOutput                     *user.Client
 }
 
 func NewClientProvider() *ClientProvider {
 	return &ClientProvider{}
+}
+
+func (c *ClientProvider) AuthClient() auth.Client {
+	c.AuthClientInvocations++
+	if c.AuthClientStub != nil {
+		return c.AuthClientStub()
+	}
+	if len(c.AuthClientOutputs) > 0 {
+		output := c.AuthClientOutputs[0]
+		c.AuthClientOutputs = c.AuthClientOutputs[1:]
+		return output
+	}
+	if c.AuthClientOutput != nil {
+		return *c.AuthClientOutput
+	}
+	panic("AuthClient has no output")
 }
 
 func (c *ClientProvider) DataSourceStructuredStore() dataSourceStoreStructured.Store {
@@ -36,27 +52,11 @@ func (c *ClientProvider) DataSourceStructuredStore() dataSourceStoreStructured.S
 	panic("DataSourceStructuredStore has no output")
 }
 
-func (c *ClientProvider) UserClient() user.Client {
-	c.UserClientInvocations++
-	if c.UserClientStub != nil {
-		return c.UserClientStub()
-	}
-	if len(c.UserClientOutputs) > 0 {
-		output := c.UserClientOutputs[0]
-		c.UserClientOutputs = c.UserClientOutputs[1:]
-		return output
-	}
-	if c.UserClientOutput != nil {
-		return *c.UserClientOutput
-	}
-	panic("UserClient has no output")
-}
-
 func (c *ClientProvider) AssertOutputsEmpty() {
+	if len(c.AuthClientOutputs) > 0 {
+		panic("AuthClientOutputs is not empty")
+	}
 	if len(c.DataSourceStructuredStoreOutputs) > 0 {
 		panic("DataSourceStructuredStoreOutputs is not empty")
-	}
-	if len(c.UserClientOutputs) > 0 {
-		panic("UserClientOutputs is not empty")
 	}
 }
