+++ user/client.go
@@ -1,9 +1,3 @@
 package user
 
-import "context"
-
-type Client interface {
-	EnsureAuthorized(ctx context.Context) error
-	EnsureAuthorizedService(ctx context.Context) error
-	EnsureAuthorizedUser(ctx context.Context, targetUserID string, permission string) (string, error)
-}
+type Client interface{}
