+++ user/client/client_test.go
@@ -3,24 +3,12 @@ package client_test
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	. "github.com/onsi/gomega/ghttp"
 
-	"context"
-	"net/http"
-
-	"github.com/tidepool-org/platform/auth"
-	authTest "github.com/tidepool-org/platform/auth/test"
 	"github.com/tidepool-org/platform/errors"
 	errorsTest "github.com/tidepool-org/platform/errors/test"
-	"github.com/tidepool-org/platform/log"
-	logTest "github.com/tidepool-org/platform/log/test"
-	"github.com/tidepool-org/platform/permission"
 	"github.com/tidepool-org/platform/platform"
-	"github.com/tidepool-org/platform/request"
-	"github.com/tidepool-org/platform/test"
 	testHttp "github.com/tidepool-org/platform/test/http"
 	userClient "github.com/tidepool-org/platform/user/client"
-	userTest "github.com/tidepool-org/platform/user/test"
 )
 
 var _ = Describe("Client", func() {
@@ -54,271 +42,6 @@ var _ = Describe("Client", func() {
 
 		It("returns success", func() {
 			Expect(userClient.New(config, authorizeAs)).ToNot(BeNil())
-		})
-	})
-
-	Context("with server and new client", func() {
-		var server *Server
-		var requestHandlers []http.HandlerFunc
-		var responseHeaders http.Header
-		var client *userClient.Client
-		var logger *logTest.Logger
-		var sessionToken string
-		var details request.Details
-		var ctx context.Context
-
-		BeforeEach(func() {
-			server = NewServer()
-			requestHandlers = nil
-			responseHeaders = http.Header{"Content-Type": []string{"application/json; charset=utf-8"}}
-			logger = logTest.NewLogger()
-			sessionToken = authTest.NewSessionToken()
-			details = request.NewDetails(request.MethodSessionToken, "", sessionToken)
-			ctx = context.Background()
-			ctx = log.NewContextWithLogger(ctx, logger)
-			ctx = auth.NewContextWithServerSessionToken(ctx, sessionToken)
-		})
-
-		JustBeforeEach(func() {
-			server.AppendHandlers(CombineHandlers(requestHandlers...))
-			var err error
-			config.Address = server.URL()
-			client, err = userClient.New(config, authorizeAs)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(client).ToNot(BeNil())
-			ctx = request.NewContextWithDetails(ctx, details)
-		})
-
-		AfterEach(func() {
-			if server != nil {
-				server.Close()
-			}
-		})
-
-		Context("EnsureAuthorized", func() {
-			Context("without server response", func() {
-				AfterEach(func() {
-					Expect(server.ReceivedRequests()).To(BeEmpty())
-				})
-
-				It("returns an error when the context is missing", func() {
-					ctx = nil
-					errorsTest.ExpectEqual(client.EnsureAuthorized(ctx), errors.New("context is missing"))
-				})
-
-				It("returns an error when the details are missing", func() {
-					ctx = request.NewContextWithDetails(ctx, nil)
-					errorsTest.ExpectEqual(client.EnsureAuthorized(ctx), request.ErrorUnauthorized())
-				})
-
-				It("returns successfully when the details are for a user", func() {
-					ctx = request.NewContextWithDetails(ctx, request.NewDetails(request.MethodSessionToken, userTest.RandomID(), sessionToken))
-					Expect(client.EnsureAuthorized(ctx)).To(Succeed())
-				})
-
-				It("returns successfully when the details are for a service", func() {
-					Expect(client.EnsureAuthorized(ctx)).To(Succeed())
-				})
-			})
-		})
-
-		Context("EnsureAuthorizedService", func() {
-			Context("without server response", func() {
-				AfterEach(func() {
-					Expect(server.ReceivedRequests()).To(BeEmpty())
-				})
-
-				It("returns an error when the context is missing", func() {
-					ctx = nil
-					errorsTest.ExpectEqual(client.EnsureAuthorizedService(ctx), errors.New("context is missing"))
-				})
-
-				It("returns an error when the details are missing", func() {
-					ctx = request.NewContextWithDetails(ctx, nil)
-					errorsTest.ExpectEqual(client.EnsureAuthorizedService(ctx), request.ErrorUnauthorized())
-				})
-
-				It("returns an error when the details are for not a service", func() {
-					ctx = request.NewContextWithDetails(ctx, request.NewDetails(request.MethodSessionToken, userTest.RandomID(), sessionToken))
-					errorsTest.ExpectEqual(client.EnsureAuthorizedService(ctx), request.ErrorUnauthorized())
-				})
-
-				It("returns successfully when the details are for a service", func() {
-					Expect(client.EnsureAuthorizedService(ctx)).To(Succeed())
-				})
-			})
-		})
-
-		Context("EnsureAuthorizedUser", func() {
-			var requestUserID string
-			var targetUserID string
-			var authorizedPermission string
-
-			BeforeEach(func() {
-				requestUserID = userTest.RandomID()
-				targetUserID = userTest.RandomID()
-				details = request.NewDetails(request.MethodSessionToken, requestUserID, sessionToken)
-				authorizedPermission = test.RandomStringFromArray([]string{permission.Write, permission.Read})
-			})
-
-			Context("without server response", func() {
-				AfterEach(func() {
-					Expect(server.ReceivedRequests()).To(BeEmpty())
-				})
-
-				It("returns an error when the context is missing", func() {
-					ctx = nil
-					userID, err := client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)
-					errorsTest.ExpectEqual(err, errors.New("context is missing"))
-					Expect(userID).To(BeEmpty())
-				})
-
-				It("returns an error when the target user id is missing", func() {
-					targetUserID = ""
-					userID, err := client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)
-					errorsTest.ExpectEqual(err, errors.New("target user id is missing"))
-					Expect(userID).To(BeEmpty())
-				})
-
-				It("returns an error when the authorized permission is missing", func() {
-					authorizedPermission = ""
-					userID, err := client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)
-					errorsTest.ExpectEqual(err, errors.New("authorized permission is missing"))
-					Expect(userID).To(BeEmpty())
-				})
-
-				It("returns an error when the details are missing", func() {
-					ctx = request.NewContextWithDetails(ctx, nil)
-					userID, err := client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)
-					errorsTest.ExpectEqual(err, request.ErrorUnauthorized())
-					Expect(userID).To(BeEmpty())
-				})
-
-				It("returns successfully when the details are for a service and authorized permission is custodian", func() {
-					ctx = request.NewContextWithDetails(ctx, request.NewDetails(request.MethodSessionToken, "", sessionToken))
-					authorizedPermission = permission.Custodian
-					Expect(client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)).To(Equal(""))
-				})
-
-				It("returns successfully when the details are for a service and authorized permission is owner", func() {
-					ctx = request.NewContextWithDetails(ctx, request.NewDetails(request.MethodSessionToken, "", sessionToken))
-					authorizedPermission = permission.Owner
-					Expect(client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)).To(Equal(""))
-				})
-
-				It("returns successfully when the details are for a service and authorized permission is upload", func() {
-					ctx = request.NewContextWithDetails(ctx, request.NewDetails(request.MethodSessionToken, "", sessionToken))
-					authorizedPermission = permission.Write
-					Expect(client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)).To(Equal(""))
-				})
-
-				It("returns successfully when the details are for a service and authorized permission is view", func() {
-					ctx = request.NewContextWithDetails(ctx, request.NewDetails(request.MethodSessionToken, "", sessionToken))
-					authorizedPermission = permission.Read
-					Expect(client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)).To(Equal(""))
-				})
-
-				It("returns an error when the details are for the target user and authorized permission is custodian", func() {
-					ctx = request.NewContextWithDetails(ctx, request.NewDetails(request.MethodSessionToken, targetUserID, sessionToken))
-					authorizedPermission = permission.Custodian
-					userID, err := client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)
-					errorsTest.ExpectEqual(err, request.ErrorUnauthorized())
-					Expect(userID).To(BeEmpty())
-				})
-
-				It("returns successfully when the details are for the target user and authorized permission is owner", func() {
-					ctx = request.NewContextWithDetails(ctx, request.NewDetails(request.MethodSessionToken, targetUserID, sessionToken))
-					authorizedPermission = permission.Owner
-					Expect(client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)).To(Equal(targetUserID))
-				})
-
-				It("returns successfully when the details are for the target user and authorized permission is upload", func() {
-					ctx = request.NewContextWithDetails(ctx, request.NewDetails(request.MethodSessionToken, targetUserID, sessionToken))
-					authorizedPermission = permission.Write
-					Expect(client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)).To(Equal(targetUserID))
-				})
-
-				It("returns successfully when the details are for the target user and authorized permission is view", func() {
-					ctx = request.NewContextWithDetails(ctx, request.NewDetails(request.MethodSessionToken, targetUserID, sessionToken))
-					authorizedPermission = permission.Read
-					Expect(client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)).To(Equal(targetUserID))
-				})
-			})
-
-			Context("with server response when the details are not for the target user", func() {
-				BeforeEach(func() {
-					requestHandlers = append(requestHandlers,
-						VerifyContentType(""),
-						VerifyHeaderKV("X-Tidepool-Session-Token", sessionToken),
-						VerifyBody(nil),
-						VerifyRequest("GET", "/access/"+targetUserID+"/"+requestUserID),
-					)
-				})
-
-				AfterEach(func() {
-					Expect(server.ReceivedRequests()).To(HaveLen(1))
-				})
-
-				Context("with an unauthenticated response", func() {
-					BeforeEach(func() {
-						requestHandlers = append(requestHandlers, RespondWith(http.StatusUnauthorized, nil, responseHeaders))
-					})
-
-					It("returns an error", func() {
-						userID, err := client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)
-						errorsTest.ExpectEqual(err, errors.New("unable to get user permissions"))
-						Expect(userID).To(Equal(""))
-					})
-				})
-
-				Context("with a not found response, which is the same as unauthorized", func() {
-					BeforeEach(func() {
-						requestHandlers = append(requestHandlers, RespondWith(http.StatusNotFound, nil, responseHeaders))
-					})
-
-					It("returns an error", func() {
-						userID, err := client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)
-						errorsTest.ExpectEqual(err, request.ErrorUnauthorized())
-						Expect(userID).To(Equal(""))
-					})
-				})
-
-				Context("with a successful response, but with no permissions", func() {
-					BeforeEach(func() {
-						requestHandlers = append(requestHandlers, RespondWith(http.StatusOK, "{}", responseHeaders))
-					})
-
-					It("returns an error", func() {
-						userID, err := client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)
-						errorsTest.ExpectEqual(err, request.ErrorUnauthorized())
-						Expect(userID).To(Equal(""))
-					})
-				})
-
-				Context("with a successful response with incorrect permissions", func() {
-					BeforeEach(func() {
-						requestHandlers = append(requestHandlers, RespondWith(http.StatusOK, `{"view": {}}`, responseHeaders))
-					})
-
-					It("returns an error", func() {
-						authorizedPermission = permission.Write
-						userID, err := client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)
-						errorsTest.ExpectEqual(err, request.ErrorUnauthorized())
-						Expect(userID).To(Equal(""))
-					})
-				})
-
-				Context("with a successful response with correct permissions", func() {
-					BeforeEach(func() {
-						requestHandlers = append(requestHandlers, RespondWith(http.StatusOK, `{"upload": {}, "view": {}}`, responseHeaders))
-					})
-
-					It("returns successfully with expected permissions", func() {
-						authorizedPermission = permission.Write
-						Expect(client.EnsureAuthorizedUser(ctx, targetUserID, authorizedPermission)).To(Equal(requestUserID))
-					})
-				})
-			})
 		})
 	})
 })
